feat: refactor SPARQL data download logic and add construct query support
